+++ test/constants/api.js
@@ -1,0 +1,5 @@
+module.exports = {
+    BASE_URL: "https://reqres.in",
+    RESTFUL_BASE_URL: "https://restful-booker.herokuapp.com",
+    FAKER_BASE_URL: "https://fakerestapi.azurewebsites.net"
+}
