+++ test/specs/deleteTest.spec.js
@@ -5,7 +5,7 @@
    you may not use this file except in compliance with the License.
    You may obtain a copy of the License at
 
-	   http://www.apache.org/licenses/LICENSE-2.0
+       http://www.apache.org/licenses/LICENSE-2.0
 
    Unless required by applicable law or agreed to in writing, software
    distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,22 +13,14 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-
 const request = require('supertest');
 const { expect } = require('chai');
+const { BASE_URL } = require('../constants/api');
 
-describe('Put API tests using supertest', () => {
-	const baseurl = 'https://reqres.in';
-	it('should successfully pass the test for post api', async () => {
-		const response = await request(baseurl)
-			.put('/api/users/2')
-			.send({ name: 'Joseph', job: 'Business Analyst' })
-			.set('Accept', 'application/json')
-			.set('Content-Type', 'application/json');
+describe('Delete API tests using supertest', () => {
+	it('should successfully pass the test for delete request', async () => {
+		const response = await request(BASE_URL).delete('/api/users/2');
 
-		expect(response.statusCode).to.be.equal(200);
-		expect(response.body.name).to.be.equal('Joseph');
-		expect(response.body.job).to.be.equal('Business Analyst');
-		expect(response.body.updatedAt).not.to.be.null;
+		expect(response.statusCode).to.be.equal(204);
 	});
 });
